delete blog with token only from the creator

+++ part4/blog-list-application/controllers/blogs.js
@@ -46,6 +46,17 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
+
+  if (!blog) {
+    return response.status(404).json({ error: "Blog not found" });
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET_KEY);
+  if (!decodedToken.id || decodedToken.id !== blog.user.toString()) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
   const result = await Blog.findByIdAndRemove(request.params.id);
   result
     ? response.status(204).end()
